has code trait
