+++ app/Models/Section.php
@@ -2,35 +2,20 @@
 
 namespace App\Models;
 
+use App\Models\Concerns\HasCode;
 use Illuminate\Database\Eloquent\Model;
 use Spatie\Translatable\HasTranslations;
 
 class Section extends Model
 {
-    use HasTranslations;
+    use HasCode, HasTranslations;
 
     public $incrementing = false;
 
     protected $translatable = ['title', 'description'];
 
-    public function code()
-    {
-        return $this->morphOne(Code::class, 'codeable');
-    }
-
     public function items()
     {
         return $this->belongsToMany(Item::class);
-    }
-
-    protected static function booted()
-    {
-        static::created(function (self $item) {
-            $item->code()->create();
-        });
-
-        static::deleting(function (self $item) {
-            $item->code()->delete();
-        });
     }
 }
